refactor(updates): isolate automatic update setting
