+++ src/renderer/src/i18n/locales/zh.json
@@ -5094,6 +5094,8 @@
           "f44299636f": "重新启动以更新（",
           "42717918f4": "安装更新（",
           "02dc082e70": "无法开始更新下载。",
+          "a73f0c19be": "Automatic updates",
+          "b8e724f0d3": "Download new updates automatically in the background. Once an update is downloaded, Orca installs it the next time you quit — your terminal sessions are never interrupted.",
           "e1a647adc5": "检查更新",
           "ceb579abaf": "检查应用程序更新并安装较新的 Orca 版本。",
           "d91ebfb87e": "当前版本：{{value0}}",
@@ -5110,9 +5112,7 @@
           "31fd7150cf": "正在检查更新...",
           "3394d1f663": "检查中",
           "d69a09b672": "启动时会自动检查更新。",
-          "7173352632": "空闲",
-          "a73f0c19be": "Automatic updates",
-          "b8e724f0d3": "Download new updates automatically in the background. Once an update is downloaded, Orca installs it the next time you quit — your terminal sessions are never interrupted."
+          "7173352632": "空闲"
         },
         "GeneralWorkspaceSettingsSection": {
           "3d538a98f7": "从工作区的“打开方式”菜单中选择可用的应用程序。",
